+++ iOS/Airbnb/Airbnb/Models/CalendarDates.swift
@@ -43,48 +43,3 @@ struct CalendarDates {
         return stay.contains(date)
     }
 }
-
-struct StayDates {
-    private var dates = [Date]()
-    
-    var checkIn: Date? {
-        if dates.count == 2 {
-            return formerOne(dates.first!, dates.last!)
-        }
-        return nil
-    }
-    
-    var checkOut: Date? {
-        if dates.count == 2 {
-            return laterOne(dates.first!, dates.last!)
-        }
-        return nil
-    }
-    
-    var oneDate: Date? {
-        if dates.count == 1 {
-            return dates.first
-        }
-        return nil
-    }
-    
-    mutating func update(to date: Date) {
-        if dates.count > 1 {
-            dates.removeFirst()
-        }
-        dates.append(date)
-    }
-    
-    func contains(_ date: Date) -> Bool {
-        guard let checkIn = checkIn, let checkOut = checkOut else { return false }
-        return (checkIn...checkOut).contains(date)
-    }
-    
-    private func formerOne(_ firstDate: Date, _ lastDate: Date) -> Date {
-        return firstDate < lastDate ? firstDate : lastDate
-    }
-    
-    private func laterOne(_ firstDate: Date, _ lastDate: Date) -> Date {
-        return firstDate > lastDate ? firstDate : lastDate
-    }
-}
